requirements.txt
python-dotenv>=1.0.1
notion-client>=2.2.1
httpx>=0.27.0
tqdm>=4.66.4
numpy>=1.26.4
sentence-transformers>=3.0.1
hnswlib>=0.8.0